Abstracted compare and league views
