+++ client/src/compare/actionTypes.js
@@ -1,0 +1,5 @@
+const exports = {
+  FETCH_SPECIFIC_PLAYERS: 'FETCH_SPECIFIC_PLAYERS',
+};
+
+export default exports;
